refactor home search data access interface to data access package
